% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.kpe}
\alias{coef.kpe}
\title{Extract the point estimate from a kpe fit.}
\usage{
\method{coef}{kpe}(object, ...)
}
\arguments{
\item{object}{A \code{"kpe"} object.}

\item{...}{Ignored.}
}
\description{
Extract the point estimate from a kpe fit.
}
