% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{confint.kpe}
\alias{confint.kpe}
\title{Confidence interval for a kpe fit.}
\usage{
\method{confint}{kpe}(object, parm, level = 0.95, ...)
}
\arguments{
\item{object}{A \code{"kpe"} object.}

\item{parm}{Ignored (only one parameter, psi).}

\item{level}{Confidence level; defaults to 0.95.}

\item{...}{Ignored.}
}
\description{
Confidence interval for a kpe fit.
}
