% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate.R
\name{.cauchy_combination}
\alias{.cauchy_combination}
\title{Cauchy combination test (Liu & Xie, JASA 2020).}
\usage{
.cauchy_combination(p_values)
}
\description{
Cauchy combination test (Liu & Xie, JASA 2020).
}
\keyword{internal}
