% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross_fit.R
\name{.popsd}
\alias{.popsd}
\title{Population (ddof=0) standard deviation, matching numpy.std default.}
\usage{
.popsd(x)
}
\description{
Population (ddof=0) standard deviation, matching numpy.std default.
}
\keyword{internal}
