% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learners_policy.R
\name{.predict_reg_family}
\alias{.predict_reg_family}
\title{Predict from a fitted regression family on a new feature matrix.}
\usage{
.predict_reg_family(family, fit, feat)
}
\description{
Predict from a fitted regression family on a new feature matrix.
}
\keyword{internal}
