% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/influence.R
\name{.predict_reward}
\alias{.predict_reward}
\title{Apply reward_model$predict with (X, A) appended.}
\usage{
.predict_reward(reward_model, X, A)
}
\description{
Apply reward_model$predict with (X, A) appended.
}
\keyword{internal}
