% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learners_propensity.R
\name{.propensity_forest}
\alias{.propensity_forest}
\title{Random-forest propensity via ranger probability forest.}
\usage{
.propensity_forest(
  seed = NULL,
  num_trees = 100L,
  rf_threads = getOption("kpe.rf_threads", 1L)
)
}
\description{
Random-forest propensity via ranger probability forest.
}
\keyword{internal}
