% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learners_propensity.R
\name{.propensity_logistic}
\alias{.propensity_logistic}
\title{Multinomial logistic propensity via glmnet (cross-validated lambda).}
\usage{
.propensity_logistic(seed = NULL)
}
\description{
Multinomial logistic propensity via glmnet (cross-validated lambda).
}
\keyword{internal}
