% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.resolve_propensity}
\alias{.resolve_propensity}
\title{Convert propensity spec into an (n, K) matrix.}
\usage{
.resolve_propensity(propensity, A, K, n)
}
\description{
Convert propensity spec into an (n, K) matrix.
}
\keyword{internal}
