% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpe.R
\name{.run_shuffles}
\alias{.run_shuffles}
\title{Run the shuffle loop with optional parallelism via `future`.}
\usage{
.run_shuffles(fn, seeds, n_cores = 1L, verbose = FALSE)
}
\description{
Run the shuffle loop with optional parallelism via `future`.
}
\keyword{internal}
