% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{kpe-package}
\alias{kpe-package}
\title{kpe: Estimating the Value of Personalized Treatment Assignment}
\description{
Point estimate, confidence interval, and one-sided p-value for the personalization effect -- the expected outcome difference between the best personalized policy and the best single treatment applied to everyone. Implements the K-Fold Personalization Estimator of Li and Brunskill (2026). Supports binary and multi-arm discrete treatments, pluggable reward, contextual-policy, and best-arm learners, and is robust to mild misspecification of the plug-in models.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/StanfordAI4HI/kpe-r}
  \item Report bugs at \url{https://github.com/StanfordAI4HI/kpe-r/issues}
}

}
\author{
\strong{Maintainer}: Emma Brunskill \email{ebrun@cs.stanford.edu} (\href{https://orcid.org/0000-0002-2200-0019}{ORCID})

Authors:
\itemize{
  \item Emma Brunskill \email{ebrun@cs.stanford.edu} (\href{https://orcid.org/0000-0002-2200-0019}{ORCID})
  \item Zhaoqi Li
}

}
\keyword{internal}
