% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.kpe}
\alias{print.kpe}
\title{Print a short summary of a kpe fit.}
\usage{
\method{print}{kpe}(x, digits = 4, ...)
}
\arguments{
\item{x}{A \code{"kpe"} object.}

\item{digits}{Number of significant digits.}

\item{...}{Ignored.}
}
\description{
Print a short summary of a kpe fit.
}
