% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.kpe}
\alias{summary.kpe}
\title{Summary of a kpe fit.}
\usage{
\method{summary}{kpe}(object, ...)
}
\arguments{
\item{object}{A \code{"kpe"} object.}

\item{...}{Ignored.}
}
\description{
Summary of a kpe fit.
}
